requirements-dev.txt
# Development-only dependencies
-r requirements.txt

# Testing
pytest>=7.4.3
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0

# Code quality
flake8>=7.0.0
black>=23.12.1
isort>=5.13.2
pylint>=3.0.3
pylint-django>=2.5.5

# Type checking
mypy>=1.8.0
django-stubs>=4.2.7

# Debugging
ipython>=8.19.0
ipdb>=0.13.13
django-debug-toolbar>=4.2.0

# Documentation
sphinx>=7.2.6
